+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/wallet/recharge/RechargePresenter.java
@@ -44,7 +44,7 @@ public class RechargePresenter extends AppBasePresenter<RechargeContract.Reposit
             mRootView.initmRechargeInstructionsPop();
             return;
         }
-        mSystemRepository.getPayStr(channel, amount).doOnSubscribe(() -> {
+        mSystemRepository.getPayStr(channel,(long) amount).doOnSubscribe(() -> {
             mRootView.configSureBtn(false);
             mRootView.showSnackLoadingMessage(mContext.getString(R.string.recharge_credentials_ing));
         }).subscribe(new BaseSubscribeForV2<PayStrBean>() {
